requirements-dev.txt
# Development and testing dependencies

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1

# Code quality
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0

# Type checking
types-requests>=2.31.0
pandas-stubs>=2.0.3.230814

# Build tools
build>=1.0.0
twine>=4.0.2

# Documentation (optional)
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
